Removed old main function and replaced with the newer version.
